fix support for outgoing TCP questions where send() does not manage to transmit the entire question in one go (should be rare) fix big-endian TCP processing of large answers, which might possibly impact intel as well
git-svn-id: svn://svn.powerdns.com/pdns/trunk/pdns@903 d19b8d6e-7fed-0310-83ef-9ca221ded41b

+++ pdns/lwres.cc
@@ -100,7 +100,7 @@ int LWRes::asyncresolve(const ComboAddress& ip, const string& domain, int type, 
       s.setNonBlocking();
       s.connect(ie);
       
-      unsigned int len=htons(vpacket.size());
+      uint16_t len=htons(vpacket.size());
       char *lenP=(char*)&len;
       const char *msgP=(const char*)&*vpacket.begin();
       string packet=string(lenP, lenP+2)+string(msgP, msgP+vpacket.size());
